osx: shut up gme warnings

+++ plugins/gme/cgme.c
@@ -312,17 +312,17 @@ cgme_add_meta (DB_playItem_t *it, const char *key, const char *value) {
     size_t len = strlen (value);
     char out[1024];
     // check for utf8 (hack)
-    if (deadbeef->junk_iconv (value, len, out, sizeof (out), "utf-8", "utf-8") >= 0) {
+    if (deadbeef->junk_iconv (value, (int)len, out, sizeof (out), "utf-8", "utf-8") >= 0) {
         deadbeef->pl_add_meta (it, key, out);
         return;
     }
 
-    if (deadbeef->junk_iconv (value, len, out, sizeof (out), "cp1252", "utf-8") >= 0) {
+    if (deadbeef->junk_iconv (value, (int)len, out, sizeof (out), "cp1252", "utf-8") >= 0) {
         deadbeef->pl_add_meta (it, key, out);
         return;
     }
 
-    if (deadbeef->junk_iconv (value, len, out, sizeof (out), "SHIFT-JIS", "utf-8") >= 0) {
+    if (deadbeef->junk_iconv (value, (int)len, out, sizeof (out), "SHIFT-JIS", "utf-8") >= 0) {
         deadbeef->pl_add_meta (it, key, out);
         return;
     }
